refactor(recipes): relocate ingredient-balance provider as consumer-owned contract
The contract now lives where it's consumed (Recipes.Application) instead
of in Yumney.Shared. The signature is unchanged — the parameter-less
GetAvailableIngredientsAsync was already trimmed in a3d9583f.

- new Recipes.Application.Interfaces.IIngredientBalanceProvider
- HttpIngredientBalanceProvider relocated to Recipes.Infrastructure/ExternalServices
- removed Shared.Common.IIngredientBalanceProvider
- Freshness enum stays in Shared.Common (a data primitive used by both
  Recipes consumers and Shopping producers — not a service interface)
- dropped a stale <see cref> in GetCookableRecipesQuery's doc summary
